djangocrm/backend/requirements.txt
django==5.2.11
celery==5.6.2
python-dotenv==1.2.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-ses==4.7.1
psycopg2-binary==2.9.11
whitenoise==6.11.0
sentry-sdk==2.53.0
django_storages==1.14.6
boto3==1.38.34

django-crum==0.7.9
disposable-email-domains>=0.0.162
redis==7.2.0
pytz==2025.2
requests==2.32.5
httpx==0.28.1
faker==40.4.0
python-dateutil==2.9.0.post0
google-auth==2.48.0

# PDF Generation
weasyprint==68.1
cairocffi==1.7.1

# Production server
gunicorn==25.1.0

# Testing
pytest==9.0.2
pytest-django==4.12.0
pytest-cov==7.0.0
hypothesis==6.131.0
hypothesis[django]